bmps/fastapi/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.1 